SCRUM-15: Get rid of scraper_with_selenium_v2.py and merge back into scraper_with_selenium.py
